moved drawing to a top level folder. Moved drawHandles.js from manipulators to drawing directory.
