+++ src/drawing/drawLinkedTextBox.js
