+++ src/tools/ArrowAnnotateTool.js
@@ -7,21 +7,21 @@ import BaseAnnotationTool from '../base/BaseAnnotationTool.js';
 import toolStyle from '../stateManagement/toolStyle.js';
 import textStyle from '../stateManagement/textStyle.js';
 import toolColors from '../stateManagement/toolColors.js';
-import drawHandles from '../manipulators/drawHandles.js';
-import drawArrow from '../util/drawArrow.js';
 import moveNewHandle from '../manipulators/moveNewHandle.js';
 import moveNewHandleTouch from '../manipulators/moveNewHandleTouch.js';
 import anyHandlesOutsideImage from '../manipulators/anyHandlesOutsideImage.js';
 import pointInsideBoundingBox from '../util/pointInsideBoundingBox.js';
-import drawLinkedTextBox from '../util/drawLinkedTextBox.js';
+import lineSegDistance from '../util/lineSegDistance.js';
 import {
   addToolState,
   removeToolState,
   getToolState
 } from '../stateManagement/toolState.js';
 import { state } from '../store/index.js';
-import lineSegDistance from '../util/lineSegDistance.js';
-import { getNewContext, draw, setShadow } from '../util/drawing.js';
+import drawLinkedTextBox from '../drawing/drawLinkedTextBox.js';
+import { getNewContext, draw, setShadow } from '../drawing/index.js';
+import drawArrow from '../drawing/drawArrow.js';
+import drawHandles from '../drawing/drawHandles.js';
 
 /**
  * @export @public @class
