+++ src/tools/WwwcTool.js
@@ -1,5 +1,5 @@
-import external from './../externalModules.js';
-import BaseTool from './../base/BaseTool.js';
+import external from '../externalModules.js';
+import BaseTool from '../base/BaseTool.js';
 
 /**
  * @export @public @class
